refactor(frontend): extract AdminPanel from page.tsx (P1.c)
Moves the admin route UI — 6 useState vars, 3 load functions, 5
handlers, 1 useMemo, ~340 lines of JSX — into
app/components/AdminPanel.tsx. page.tsx drops 500 lines.

Key behavioral wins:
- Admin state is only allocated when showAdmin is true; navigating
  away unmounts the panel and its fetchers die naturally, closing a
  class of poll-leak bugs.
- Global setStatusMessage/setErrorMessage stay in page.tsx; AdminPanel
  calls them via onStatus/onError props. libraryEntries still flows
  down as a prop because it is also owned by page.tsx for other panels.
- Load-on-mount replaces the showAdmin-gated effect that fired on the
  parent's render cycle.

Also adds explicit 'import React' to MermaidDiagram for consistency;
vitest's jsdom runtime fails on JSX without it in new component files
(caught by the admin overlay regression test).

Verified: frontend 46/46 pass (stable across 5 full-suite reruns),
production build succeeds, /admin route renders identical UI.
